Update recent games in Tray Icon, backend events, some refactor
