+++ src/backend/main.ts
@@ -1,3 +1,4 @@
+import { EventEmitter } from 'node:events'
 import { initImagesCache } from './images_cache'
 import { downloadAntiCheatData } from './anticheat/utils'
 import {
@@ -11,15 +12,13 @@ import * as path from 'path'
 import {
   BrowserWindow,
   Menu,
-  Tray,
   app,
   dialog,
   ipcMain,
   powerSaveBlocker,
   protocol,
   screen,
-  clipboard,
-  nativeImage
+  clipboard
 } from 'electron'
 import 'backend/updater'
 import { autoUpdater } from 'electron-updater'
@@ -81,8 +80,6 @@ import {
   heroicGithubURL,
   userHome,
   icon,
-  iconDark,
-  iconLight,
   installed,
   kofiPage,
   epicLoginUrl,
@@ -123,7 +120,7 @@ import {
   runOnceWhenOnline
 } from './online_monitor'
 import { showDialogBoxModalAuto } from './dialog/dialog'
-import { addRecentGame, getRecentGames } from './recent_games'
+import { addRecentGame } from './recent_games'
 import {
   addNewApp,
   appLogFileLocation,
@@ -138,11 +135,14 @@ import {
 import { callAbortController } from './utils/aborthandler/aborthandler'
 import { getDefaultSavePath } from './save_sync'
 import si from 'systeminformation'
+import { initTrayIcon } from './tray_icon'
 
 const { showOpenDialog } = dialog
 const isWindows = platform() === 'win32'
 
 let mainWindow: BrowserWindow
+
+export const backendEvents = new EventEmitter()
 
 async function createWindow(): Promise<BrowserWindow> {
   configStore.set('userHome', userHome)
@@ -278,55 +278,6 @@ async function createWindow(): Promise<BrowserWindow> {
 const gotTheLock = app.requestSingleInstanceLock()
 let openUrlArgument = ''
 
-const contextMenu = async () => {
-  const recentsMenu = (await getRecentGames({ limited: true })).map((game) => {
-    return {
-      click: function () {
-        handleProtocol(mainWindow, [`heroic://launch/${game.appName}`])
-      },
-      label: game.title
-    }
-  })
-
-  return Menu.buildFromTemplate([
-    ...recentsMenu,
-    { type: 'separator' },
-    {
-      click: function () {
-        mainWindow.show()
-      },
-      label: i18next.t('tray.show')
-    },
-    {
-      click: function () {
-        showAboutWindow()
-      },
-      label: i18next.t('tray.about', 'About')
-    },
-    {
-      accelerator: process.platform === 'darwin' ? 'Cmd+R' : 'Ctrl+R',
-      click: function () {
-        mainWindow.reload()
-      },
-      label: i18next.t('tray.reload', 'Reload')
-    },
-    {
-      label: 'Debug',
-      accelerator: process.platform === 'darwin' ? 'Alt+Cmd+I' : 'Ctrl+Shift+I',
-      click: () => {
-        mainWindow.webContents.openDevTools()
-      }
-    },
-    {
-      click: function () {
-        handleExit(mainWindow)
-      },
-      label: i18next.t('tray.quit', 'Quit'),
-      accelerator: process.platform === 'darwin' ? 'Cmd+Q' : 'Ctrl+Q'
-    }
-  ])
-}
-
 const processZoomForScreen = (zoomFactor: number) => {
   const screenSize = screen.getPrimaryDisplay().workAreaSize.width
   if (screenSize < 1200) {
@@ -374,7 +325,7 @@ if (!gotTheLock) {
     // TODO: Remove this after a couple of stable releases
     // Affects only current users, not new installs
     const settings = await GlobalConfig.get().getSettings()
-    const { language, darkTrayIcon } = settings
+    const { language } = settings
     const currentConfigStore = configStore.get('settings', {}) as AppSettings
     if (!currentConfigStore.defaultInstallPath) {
       configStore.set('settings', settings)
@@ -480,55 +431,17 @@ if (!gotTheLock) {
       mainWindow.webContents.setZoomFactor(processZoomForScreen(zoomFactor))
     }, 200)
 
-    const iconSizesByPlatform = {
-      darwin: {
-        width: 20,
-        height: 20
-      },
-      linux: {
-        width: 32,
-        height: 32
-      },
-      win32: {
-        width: 32,
-        height: 32
-      }
-    }
-
-    const trayIcon = nativeImage
-      .createFromPath(darkTrayIcon ? iconDark : iconLight)
-      .resize(iconSizesByPlatform[process.platform])
-
-    const appIcon = new Tray(trayIcon)
-
-    appIcon.on('double-click', () => {
-      mainWindow.show()
-    })
-
-    appIcon.setContextMenu(await contextMenu())
-    appIcon.setToolTip('Heroic')
     ipcMain.on('changeLanguage', async (event, language) => {
       logInfo(['Changing Language to:', language], {
         prefix: LogPrefix.Backend
       })
       await i18next.changeLanguage(language)
       gameInfoStore.clear()
-      appIcon.setContextMenu(await contextMenu())
-    })
-
-    ipcMain.addListener('changeTrayColor', () => {
-      logInfo('Changing Tray icon Color...', { prefix: LogPrefix.Backend })
-      setTimeout(async () => {
-        const { darkTrayIcon } = await GlobalConfig.get().getSettings()
-        const trayIcon = nativeImage
-          .createFromPath(darkTrayIcon ? iconDark : iconLight)
-          .resize(iconSizesByPlatform[process.platform])
-        appIcon.setImage(trayIcon)
-        appIcon.setContextMenu(await contextMenu())
-      }, 500)
     })
 
     downloadAntiCheatData()
+
+    initTrayIcon(mainWindow)
 
     return
   })
